+++ package.json
@@ -13,6 +13,7 @@
     "@mui/icons-material": "^6.4.2",
     "@mui/material": "^6.4.2",
     "@react-native-async-storage/async-storage": "^2.1.0",
+    "@react-native-community/blur": "^4.4.1",
     "@react-native-picker/picker": "2.9.0",
     "@react-navigation/drawer": "^7.1.1",
     "@react-navigation/native": "^7.0.14",
@@ -21,8 +22,8 @@
     "expo": "^52.0.25",
     "expo-camera": "^16.0.17",
     "expo-constants": "~17.0.3",
-    "expo-image-picker": "^16.0.6",
     "expo-env": "^1.1.1",
+    "expo-image-picker": "^16.0.6",
     "expo-linking": "~7.0.4",
     "expo-router": "~4.0.16",
     "expo-status-bar": "~2.0.1",
@@ -42,7 +43,8 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-snap-carousel": "^1.3.1",
-    "react-native-vector-icons": "^10.2.0"
+    "react-native-vector-icons": "^10.2.0",
+    "expo-blur": "~14.0.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
